refactored ``propose`` function out to ``sampler.py``
- also moved test
