+++ tests/test_sampler.py
@@ -3,11 +3,12 @@ import pytest
 from hypothesis import given, settings
 from hypothesis import strategies as st
 
-from jax_unirep.sampler import is_accepted
+from jax_unirep.sampler import is_accepted, propose
+from jax_unirep.utils import proposal_valid_letters
 
 
 @given(st.data())
-@settings(deadline=None, max_examples=100)
+# @settings(deadline=None, max_examples=100)
 def test_is_accepted(data):
     best = data.draw(
         st.floats(
@@ -17,3 +18,17 @@ def test_is_accepted(data):
     candidate = data.draw(st.floats(allow_nan=False, allow_infinity=False))
     accept = is_accepted(best=best, candidate=candidate)
     assert isinstance(accept, bool)
+
+
+@given(seq=st.text(alphabet=proposal_valid_letters, min_size=1))
+def test_propose_string(seq):
+    new_seq = propose(seq)
+    assert new_seq != seq
+    assert len(new_seq) == len(seq)
+
+    # Check that only one position is different
+    different_positions = []
+    for i, (l1, l2) in enumerate(zip(seq, new_seq)):
+        if l1 != l2:
+            different_positions.append(i)
+    assert len(different_positions) == 1
